requirements.txt
Django==2.2.13
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
mysql==0.0.2
mysqlclient==1.4.6
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
